Clean up api.
