--- imports/api/events/README.md
@@ -1,13 +1,0 @@
-# Welcome to the eventbus!
-In Grow-IoT, we have things (a.k.a. event emitters) and an eventbus api for working with events!
-
-Unfortunately, this API needs work. Some things we might consider:
-
-+ [Time series support](https://github.com/CommonGarden/Grow-IoT/issues/211)
-+ Connecting to external services or devices
-+ Adding support for rules
-+ REST endpoints
-+ [MQTT support?](https://github.com/CommonGarden/Grow-IoT/issues/224)
-+ COAP?
-+ ???
-
